Add the other 2 motors and connect motors via MotorControllerGroup.
This renames the existing motors from 'm_leftMotor' to 'm_leftMotor1'
(and m_rightMotor to m_rightMotor1) and then adds two new motors
m_leftMotor2 and m_rightMotor2.

Then, we add those two motors to a MotorControllerGroup.
MotorControllerGroup is just what it sounds like.  Its a group of
Motors that will be controlled as if they are one motor.