Re-apply local changes to SwerveSubsystem.java except for one
We had a local change getTargetSpeeds that looked like this:

    getHeading().getRadians()
       * (RobotBase.isSimulation() || DriverStation.isAutonomous() ? 1 : -1),

Rather than just

    getHeading().getRadians()

Put the rest of the changes back in as they were.

getAutonomousCommand no longer takes a boolean to reset the odometry.
That changed in YAGSL-Example PathPlanner when it changed from taking
a path to a "Path" to using a path to a "Auton".